+++ src/main/java/com/example/petShelter/service/SheltersService.java
@@ -10,11 +10,13 @@ import org.springframework.stereotype.Service;
 
 import java.util.Collection;
 import java.util.Optional;
+
 /**
  * The class consists of logic of the project, which has the methods  to work with "Shelters" entity
  *
  * @author Khilola Kushbakova
  */
+
 @Service
 public class SheltersService {
 
@@ -26,6 +28,7 @@ public class SheltersService {
     public SheltersService(SheltersRepository sheltersRepository) {
         this.sheltersRepository = sheltersRepository;
     }
+
     /**
      * Method to list all shelters.
      *
@@ -36,18 +39,30 @@ public class SheltersService {
         Collection<Shelters> shelters = sheltersRepository.findAll();
         if (shelters == null) {
             log.error("There is no any shelter ");
+            throw new RuntimeException();
         }
         if (shelters.isEmpty()) {
             log.warn("No shelters found, add a new shelter");
+            throw new RuntimeException();
         }
 
         return shelters;
     }
 
+    /**
+     * Method to add a shelter
+     *
+     * @return created shelter or throw exception if the shelter id is present
+     */
     public Shelters addShelter(Shelters shelter) {
-        log.info("Was invoked method for addShelter");
-        return sheltersRepository.save(shelter);
+        if (sheltersRepository.existsById(shelter.getId())) {
+            throw new RuntimeException();
+        } else {
+            log.info("Was invoked method for addShelter");
+            return sheltersRepository.save(shelter);
+        }
     }
+
     /**
      * Method to find a shelter by its identifier.
      *
@@ -63,20 +78,22 @@ public class SheltersService {
 
         return shelter;
     }
+
     /**
      * Removes a shelter with the given shelterId from the database.
      *
      * @param shelterId The unique identifier of the shelter to be removed
      */
     public void removeShelter(long shelterId) {
-        Shelters shelter = sheltersRepository.findById(shelterId).orElse(null);
-        sheltersRepository.deleteById(shelterId);
-        log.info("Was invoked method for removeShelter");
-        if (shelter == null) {
+        if (sheltersRepository.existsById(shelterId)) {
+            sheltersRepository.deleteById(shelterId);
+            log.info("Was invoked method for removeShelter");
+        } else {
             log.error("There is no shelter with id = {}", shelterId);
-
+            throw new RuntimeException();
         }
     }
+
     /**
      * Retrieves the contact information of a shelter by its unique identifier.
      *
@@ -88,12 +105,14 @@ public class SheltersService {
         return sheltersRepository.getReferenceById(shelterId).getContact();
 
     }
+
     public String showAddress(long shelterId) {
         log.info("Was invoked method for showAddress");
 
         final Optional<Shelters> byIdOptionalShelters = sheltersRepository.findById(shelterId);
         return byIdOptionalShelters.isPresent() ? byIdOptionalShelters.get().getAddress() : null;
     }
+
     /**
      * This method generates a string representation of a location based on the provided latitude and longitude.
      *
@@ -105,10 +124,13 @@ public class SheltersService {
         String location = "Latitude: " + latitude + ", Longitude: " + longitude;
         return location;
     }
+
     public String showSecurityNumber(long shelterId) {
         log.info("Was invoked method for giveSecurityNumber");
         return sheltersRepository.getReferenceById(shelterId).getSecurityContact();
     }
+
+
     /**
      * This method is used to change the information of a shelter in the database.
      *
@@ -119,7 +141,9 @@ public class SheltersService {
     public Shelters changeShelterInfo(Shelters shelter) {
         log.info("Was invoked method for  changeShelterInfo");
         return sheltersRepository.save(shelter);
+
     }
+
     public String showAnimalInfoById(long id) {
         log.info("Was invoked method for  showAnimalInfoById");
         final Optional<Shelters> byIdOptionalShelters = sheltersRepository.findById(id);
@@ -128,16 +152,19 @@ public class SheltersService {
                 + shelters.getAddress() + " "
                 + shelters.getContact()).orElse(null);
     }
+
     public String showSchedule(long id) {
         log.info("Was invoked method for showSchedule");
         final Optional<Shelters> byIdOptionalShelters = sheltersRepository.findById(id);
         return byIdOptionalShelters.map(Shelters::getWorkingHours).orElse(null);
     }
+
     public String showSecurityContact(long id) {
         log.info("Was invoked method for  showSecurityContact");
         final Optional<Shelters> byIdOptionalShelters = sheltersRepository.findById(id);
         return byIdOptionalShelters.map(Shelters::getSecurityContact).orElse(null);
     }
+
     public String showShelterRules(long id) {
         log.info("Was invoked method for showShelterRules");
         final Optional<Shelters> byIdOptionalShelters = sheltersRepository.findById(id);
